requirements.txt
django-qrcode==0.3
PySimpleGUI==4.41.2
selenium==3.141.0
urllib3==1.26.4
